# specs/001-lottery-sandbox/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Multi-Level Agent Lottery Sandbox API
  description: |
    RESTful API for Multi-Level Agent Lottery Sandbox System supporting Malaysian and Singapore lottery providers.

    **Key Features:**
    - Unlimited agent hierarchy with configurable commission
    - Weekly limit-based betting system (Monday reset)
    - Multiple service providers (Magnum, Sports Toto, Damacai, Singapore Pools)
    - Game types: 3D, 4D, 5D, 6D
    - Bet types: BIG, SMALL, iBox (MVP)
    - Third-party API integration with manual fallback
    - Role-based access control (ADMIN, MODERATOR, AGENT)

  version: 1.0.0
  contact:
    name: API Support
    email: support@lottery-sandbox.com
servers:
  - url: https://api.lottery-sandbox.com/api/v1
    description: Production server
  - url: https://staging-api.lottery-sandbox.com/api/v1
    description: Staging server
  - url: http://localhost:3000/api/v1
    description: Local development server
...
security:
  - BearerAuth: []
paths:
...
  /providers:
    get:
      tags:
        - Service Providers
      summary: List all service providers
      description: Get all configured lottery service providers (visible to all users)
      operationId: listProviders
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: country
          in: query
          schema:
            type: string
            enum: [MY, SG]
          description: Filter by country code
        - name: isActive
          in: query
          schema:
            type: boolean
          description: Filter by active status
      responses:
        '200':
          description: Providers retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ServiceProviderDto'
                  pagination:
                    $ref: '#/components/schemas/PaginationMeta'
              example:
                data:
                  - id: "clx1a2b3c4d5e6f7g8h9"
                    code: "M"
                    name: "Magnum 4D"
                    country: "MY"
                    isActive: true
                    availableGames: ["4D"]
                    betTypes: ["BIG", "SMALL", "IBOX"]
                    drawSchedule:
                      days: [0, 3, 6]
                      time: "19:00"
                      timezone: "Asia/Kuala_Lumpur"
                    createdAt: "2025-01-15T08:00:00Z"
                    updatedAt: "2025-01-15T08:00:00Z"
                pagination:
                  total: 4
                  page: 1
                  limit: 20
                  totalPages: 1
        '401':
          $ref: '#/components/responses/UnauthorizedError'

    post:
      tags:
        - Service Providers
      summary: Create service provider
      description: Create new lottery service provider (ADMIN only)
      operationId: createProvider
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateServiceProviderDto'
            example:
              code: "M"
              name: "Magnum 4D"
              country: "MY"
              availableGames: ["4D"]
              betTypes: ["BIG", "SMALL", "IBOX"]
              drawSchedule:
                days: [0, 3, 6]
                time: "19:00"
                timezone: "Asia/Kuala_Lumpur"
              apiConfig:
                provider: "magayo"
                apiKey: "{{MAGAYO_API_KEY}}"
                endpoint: "https://api.magayo.com/lottery/results"
      responses:
        '201':
          description: Provider created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceProviderDto'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '422':
          $ref: '#/components/responses/ValidationError'
  /providers/{providerId}:
    get:
      tags:
        - Service Providers
      summary: Get service provider details
      operationId: getProvider
      parameters:
        - $ref: '#/components/parameters/ProviderIdParam'
      responses:
        '200':
          description: Provider retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceProviderDto'
        '404':
          $ref: '#/components/responses/NotFoundError'

    patch:
      tags:
        - Service Providers
      summary: Update service provider
      description: Update service provider configuration (ADMIN only)
      operationId: updateProvider
      parameters:
        - $ref: '#/components/parameters/ProviderIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateServiceProviderDto'
      responses:
        '200':
          description: Provider updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceProviderDto'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'

    delete:
      tags:
        - Service Providers
      summary: Delete service provider
      description: Soft delete service provider (ADMIN only)
      operationId: deleteProvider
      parameters:
        - $ref: '#/components/parameters/ProviderIdParam'
      responses:
        '204':
          description: Provider deleted successfully
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'

  # ============================================================================
  # AGENTS
  # ============================================================================
...
  /results:
    get:
      tags:
        - Draw Results
      summary: List draw results
      operationId: listResults
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: providerId
          in: query
          schema:
            type: string
          description: Filter by provider ID
        - name: gameType
          in: query
          schema:
            type: string
            enum: [3D, 4D, 5D, 6D]
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Results retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/DrawResultDto'
                  pagination:
                    $ref: '#/components/schemas/PaginationMeta'

    post:
      tags:
        - Draw Results
      summary: Create draw result manually
      description: Manually enter draw result (ADMIN only, used as API fallback)
      operationId: createResult
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDrawResultDto'
      responses:
        '201':
          description: Result created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DrawResultDto'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /results/{resultId}:
    get:
      tags:
        - Draw Results
      summary: Get draw result details
      operationId: getResult
      parameters:
        - $ref: '#/components/parameters/ResultIdParam'
      responses:
        '200':
          description: Result retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DrawResultDto'
  /results/sync:
    post:
      tags:
        - Draw Results
      summary: Sync results from third-party API
      description: Trigger manual sync of lottery results from third-party provider (ADMIN only)
      operationId: syncResults
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                providerId:
                  type: string
                  description: Provider ID to sync (optional, syncs all if omitted)
                drawDate:
                  type: string
                  format: date
                  description: Draw date to sync (optional, syncs latest if omitted)
      responses:
        '200':
          description: Sync completed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Synced 4 draw results"
                  syncedResults:
                    type: array
                    items:
                      $ref: '#/components/schemas/DrawResultDto'
        '403':
          $ref: '#/components/responses/ForbiddenError'

  # ============================================================================
  # COMMISSIONS
  # ============================================================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT access token (expires in 15 minutes)

        Example:
        ```
        Authorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        ```
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Page number (1-indexed)
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Items per page
    ProviderIdParam:
      name: providerId
      in: path
      required: true
      schema:
        type: string
      description: Service provider ID (CUID)
...
    ResultIdParam:
      name: resultId
      in: path
      required: true
      schema:
        type: integer
      description: Draw result ID
  schemas:
...
    ServiceProviderDto:
      type: object
      properties:
        id:
          type: string
          example: "clx1a2b3c4d5e6f7g8h9"
        code:
          type: string
          example: "M"
        name:
          type: string
          example: "Magnum 4D"
        country:
          type: string
          enum: [MY, SG]
          example: "MY"
        isActive:
          type: boolean
          example: true
        availableGames:
          type: array
          items:
            type: string
            enum: [3D, 4D, 5D, 6D]
          example: ["4D"]
        betTypes:
          type: array
          items:
            type: string
            enum: [BIG, SMALL, IBOX]
          example: ["BIG", "SMALL", "IBOX"]
        drawSchedule:
          type: object
          properties:
            days:
              type: array
              items:
                type: integer
                minimum: 0
                maximum: 6
              description: Days of week (0=Sunday, 6=Saturday)
              example: [0, 3, 6]
            time:
              type: string
              pattern: '^([0-1][0-9]|2[0-3]):[0-5][0-9]$'
              example: "19:00"
            timezone:
              type: string
              example: "Asia/Kuala_Lumpur"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateServiceProviderDto:
      type: object
      required:
        - code
        - name
        - country
        - availableGames
        - betTypes
        - drawSchedule
      properties:
        code:
          type: string
          minLength: 1
          maxLength: 10
        name:
          type: string
          minLength: 1
          maxLength: 100
        country:
          type: string
          enum: [MY, SG]
        availableGames:
          type: array
          items:
            type: string
            enum: [3D, 4D, 5D, 6D]
          minItems: 1
        betTypes:
          type: array
          items:
            type: string
            enum: [BIG, SMALL, IBOX]
          minItems: 1
        drawSchedule:
          type: object
          required:
            - days
            - time
            - timezone
          properties:
            days:
              type: array
              items:
                type: integer
                minimum: 0
                maximum: 6
            time:
              type: string
            timezone:
              type: string
        apiConfig:
          type: object
          properties:
            provider:
              type: string
            apiKey:
              type: string
            endpoint:
              type: string
    UpdateServiceProviderDto:
      type: object
      properties:
        name:
          type: string
        isActive:
          type: boolean
        availableGames:
          type: array
          items:
            type: string
            enum: [3D, 4D, 5D, 6D]
        betTypes:
          type: array
          items:
            type: string
            enum: [BIG, SMALL, IBOX]
        drawSchedule:
          type: object
        apiConfig:
          type: object

    # ========================================================================
    # BET SCHEMAS
    # ========================================================================
...
    DrawResultDto:
      type: object
      properties:
        id:
          type: integer
          example: 1001
        providerId:
          type: string
          example: "clx1a2b3c4d5e6f7g8h9"
        gameType:
          type: string
          enum: [3D, 4D, 5D, 6D]
          example: "4D"
        drawDate:
          type: string
          format: date
          example: "2025-01-18"
        drawTime:
          type: string
          format: time
          example: "19:00:00"
        winningNumbers:
          type: object
          description: Winning numbers by prize category
          example:
            first: "1234"
            second: "5678"
            third: "9012"
            special: ["3456", "7890", "1122", "3344", "5566", "7788", "9900", "1230", "4560", "7891"]
            consolation: ["2345", "6789", "0123", "4567", "8901", "2346", "6790", "0124", "4568", "8902"]
        source:
          type: string
          enum: [API, MANUAL]
          example: "API"
        syncedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
    CreateDrawResultDto:
      type: object
      required:
        - providerId
        - gameType
        - drawDate
        - drawTime
        - winningNumbers
      properties:
        providerId:
          type: string
        gameType:
          type: string
          enum: [3D, 4D, 5D, 6D]
        drawDate:
          type: string
          format: date
        drawTime:
          type: string
          format: time
        winningNumbers:
          type: object

    # ========================================================================
    # COMMISSION SCHEMAS
    # ========================================================================
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        totalPages:
          type: integer
          example: 5
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "BAD_REQUEST"
        message:
          type: string
          example: "Insufficient weekly limit"
        statusCode:
          type: integer
          example: 400
        details:
          type: object
          nullable: true
  responses:
    UnauthorizedError:
      description: Authentication failed or token expired
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "UNAUTHORIZED"
            message: "Invalid or expired token"
            statusCode: 401
    ForbiddenError:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "FORBIDDEN"
            message: "Insufficient permissions to perform this action"
            statusCode: 403
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "NOT_FOUND"
            message: "Resource not found"
            statusCode: 404
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "VALIDATION_ERROR"
            message: "Request validation failed"
            statusCode: 422
            details:
              - field: "betAmount"
                message: "betAmount must be a positive number"